+++ src/lib/lforms/lhc-form-utils.js
@@ -202,6 +202,60 @@ const FormUtils = {
 
 
   /**
+   * Makes a Questionnaire/$validate call with the given Questionnaire and base FHIR server.
+   * @param questionnaire a FHIR Questionnaire resource to be validated.
+   * @param fhirServerBase the base URL of the FHIR server to which the $validate call should be made.
+   * @return a Promise that resolves to null if the Questionnaire is valid, or an error message if not valid.
+   * If multiple errors are returned from server, they will be concatenated into a single string with " | " as separator.
+   */
+  validateQuestionnaireOnFHIRServer: function(questionnaire, fhirServerBase) {
+    if (!questionnaire || questionnaire.resourceType !== 'Questionnaire') {
+      throw new Error('The provided resource is not a Questionnaire.');
+    }
+    if (!fhirServerBase) {
+      throw new Error('FHIR server base URL is required for validating the Questionnaire.');
+    }
+    let fullUrl;
+    try {
+      // Add a trailing slash if it doesn't have one, so the path in fhirServerBase would not be lost.
+      if (!fhirServerBase.endsWith('/')) {
+        fhirServerBase = fhirServerBase + '/';
+      }
+      fullUrl = new URL('Questionnaire/$validate', fhirServerBase);
+    } catch (error) {
+      throw new Error('The URL was malformed!');
+    }
+    return fetch(fullUrl.href, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(questionnaire)
+    })
+      .then(function (response) {
+        if (!response.ok) {
+          throw new Error(`Server returned ${response.status}: ${response.statusText}`);
+        }
+        return response.json();
+      })
+      .then(function (parsedJSON) {
+        let rtn = null;
+        if (parsedJSON.resourceType === "OperationOutcome") {
+          const errorOrFatal = parsedJSON.issue?.filter(item => item.severity === "error" || item.severity === "fatal");
+          if (errorOrFatal && errorOrFatal.length) {
+            rtn = errorOrFatal.map(e => e.diagnostics).join(' | ');
+          }
+        }
+        return rtn;
+      })
+      .catch(function (error) {
+        // Catch network failures and http errors like 404, 500.
+        throw new Error('Fetch error: ' + error.message || error);
+      });
+  },
+
+
+  /**
    * Convert LForms data into a FHIR resource
    * @param resourceType a FHIR resource type. it currently supports "DiagnosticReport",
    * "Questionnaire" (both standard Questionnaire and SDC Questionnaire profile)